style: satisfy ruff lint checks
The first CI run surfaced ruff 0.16 findings the older local ruff missed (UP037, I001, RUF046, RUF059, PLC0206). All are style-level with no behaviour change.

+++ scripts/validate_fukuchi.py
@@ -65,7 +65,7 @@ def analyze_trial(mkr: pd.DataFrame, grf: pd.DataFrame) -> list[tuple]:
     keys, com, ct, cp, mm = [], [], [], [], []
     for i in range(nm):
         FUNNEL["frames"] += 1
-        j = int(round(i * (ng - 1) / (nm - 1)))
+        j = round(i * (ng - 1) / (nm - 1))
         loaded = [p for p in range(5) if fv[p][j] > FV_THRESH]
         if len(loaded) != 1:
             continue
